pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wao"
version = "0.1.0"
description = "Web Automatic Operation CLI (skeleton)"
requires-python = ">=3.10"
dependencies = ["jsonschema>=4.23.0", "python-dotenv>=1.0.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
